+++ website/documentation-site.ts
@@ -21,6 +21,12 @@ export const DOCUMENTATION_DEFAULT_VERSION = 'latest';
 
 export const DOCUMENTATION_VERSIONS = [
   'latest',
+  '0.255.0',
+  '0.254.0',
+  '0.253.0',
+  '0.252.0',
+  '0.251.0',
+  '0.250.0',
   '0.249.0',
   '0.248.0',
   '0.247.0',
@@ -35,12 +41,6 @@ export const DOCUMENTATION_VERSIONS = [
   '0.238.0',
   '0.237.0',
   '0.236.0',
-  '0.235.0',
-  '0.234.0',
-  '0.233.0',
-  '0.232.0',
-  '0.231.0',
-  '0.230.0',
   '0.38.0',
   '0.1.0',
 ] as const;